backend/requirements.txt
Django>=5.0
#django-cors-headers
djangorestframework>=3.14
djangorestframework-simplejwt
django-cors-headers==4.3.1
python-dotenv==1.0.0
gunicorn
django-cors-headers
drf-yasg>=1.21.7
